RFC: add vlc_player_t (input manager)
Hello,

First, I decided to name the future input manager/controller vlc_player_t.
For me it is really a player, really like the libvlc one
(libvlc_media_player_t). Note than the libvlc_media_player_t will use this
vlc_player_t in the future.

This commit is an RFC, it's not yet tested (I will add a test), or used.

I discussed a lot with Romain, that is working on the next playlist, to provide
an API that will be easy to use for him.

The main vlc_player_t will be created and owned by the main playlist. Sub
players and playlist can still be created/used from any modules (but I don't
know the real use case for that). The playlist will take care of setting the
current and next medias via vlc_player_SetCurrentMedia() and via the
vlc_player_owner_cbs.get_next_media callback. The API was thought for gapless,
even if it still not implemented.

Any interface or controls modules will be able to retrieve the main player
instance via the main playlist instance. They will be able to register for
playback events (the same than input_thread_t events for now), get the full
state of the player (length, paused, tracks, programs...) and control it.

The lock of the player will be exposed by the API, and each function must be
called while the player is locked. This allows finer controls of the player,
you can now do actions in function of the player state in a non racy way.
Player functions can also be called from player event callbacks.

All controls of the player are asynchronous, even vlc_player_Stop(). I saw too
many hacks in VLC ports for stopping the player, they generally spawn a
detached thread that will stop it. Indeed, input_Close() is blocking, it wait
for the input thread to be terminated. This should be instantaneous for basic
remote accesses that handle vlc_interrupt. But what do we do for remote
accesses that need a teardown of local accesses (yes, it's really easy to block
on a USB drive) ?

My proposal is the destructor struct from player.c. When the user need to stop
a playback, the player will prevent input events forwarding, call input_Stop()
and won't wait for the input thread termination. Instead, when the
INPUT_EVENT_DEAD event will be received, the input_thread_t instance will be
passed to the destructor thread that will join it.

One problem: vlc_player_Delete() will join this destructor thread and then wait
for all input threads termination (but I don't know what else we can do without
leaking).

Still not handled in this patch
 - Test
 - Seeking
 - Get position/time
 - Some state not handling (missing getters)
 - gapless
 - aout/vout handling. Do we expose them or add API to control the
   volume/select the audio device for example.

What do you think about this patch so far ?

Regards,
Thomas