Introduce media source and media tree API
Add an API to manage "services discovery" out of the playlist.

A "media source provider", associated to the libvlc instance, allows to
retrieve media sources (each associated to a services discovery module).

Requesting a services discovery that is not open will automatically open
it. If several "clients" request the same media source (i.e. by
requesting the same name), they will receive the same (refcounted) media
source instance.

As soon as a media source is released by all its clients, the associated
services discovery is closed.

Each media source holds a media tree, independant of the playlist, used
to store the input items detected by the services discovery. Clients may
listen to the tree to be notified of changes.

A function to preparse an item is exposed on media tree. If subitems are
detected during the preparsing, it updates the media tree accordingly.