Use new media source API from the playlist
A future UI would use media trees instead of the playlist (currently, in
PLModel for Qt).

As an incremental step, in order not to break the current UI, make the
playlist a client of the new API, so that every changes applied to the
media trees are reflected to the playlist. That way, the media detected
by services discovery are still displayed in the UI.

This is a transitional implementation to be able to update the UI (Qt
and MacOS) separately.