vlc_arrays: refactor ARRAY_FOREACH()
Replace:

    FOREACH_ARRAY(item, array)
       // ...
    FOREACH_END()

by:

    ARRAY_FOREACH(item, array)
    {
        // ...
    }

The new implementation uses a variable name which depends on "item" to
avoid variable shadowing warnings for nested loops (even if this
specific shadowing is harmless). See previous discussion:
<https://mailman.videolan.org/pipermail/vlc-devel/2018-June/119489.html>

As a consequence, it prevents to declare the variable directly.
Concretely, it is not possible to write:

    ARRAY_FOREACH(void *item, array)

Instead, use:

    void *item;
    ARRAY_FOREACH(item, array);