preparser: use fetcher callbacks
After a preparsing, the preparser may request to fetch art, and trigger
the "preparse end" event only after the art have been fetched.

For that purpose, the fetcher and the preparser were highly coupled: the
fetcher was aware of a "preparse status", and was responsible to notify
"preparse ended" on the input item when necessary.

In order to make the fetcher more independant of the preparser, make the
preparser trigger the "preparse ended" event when it is notified of the
"fetch ended" event.

This also paves the way to remove preparsing events on the input item
directly.