core: playlist: new playlist API
Add a new playlist API.

A playlist contains a simple list of items, and owns a player.

Callbacks are exposed so that clients (UI) are notified when items are
updated, insert, moved or removed. The playlist is responsible for the
playback order and repeat mode, and manages a cursor to the "current"
item.