playlist: remove lua search
The lua API exposed a function to search in (actually, filter) the
playlist. The filtering was directly applied on the core playlist,
so every playlist clients shared the same filter.

The new playlist does not filter. Instead, each client could create its
own filtered view of the core playlist.

We do not want to invest time to implement this for the lua API for now,
so remove the search feature of the "cli" interface.