qt: playlist: refactor "move items"
Moving a set of items to a given target is ambiguous: the target index
may be expressed either _before_ or _after_ the move is actually
applied.

                pre-target
   move +----------+                       target
        |          |                         |
    |-------|      v                         v
  A [B  C  D] E  F  G      ------>   A  E  F [B  C  D] G
  0  1  2  3  4  5  6                0  1  2  3  4  5  6

The core playlist expects the target to be the position _after_ the
move.

In the UI, we need both:
 - during a drag&drop, we naturally know the position of the "drop",
   which is the target _before_ the move;
 - during a move using key navigation, we shift the items by a given
   amount (typically 1), so we know the expected target _after_ the
   move.

To express a move request in a "natural" way, expose two separate
functions to move items in Qt, one accepting the pre-target and another
the post-target.