qt: playlist: update drag&drop count on drag
To display the number of items selected during a drag&drop, the text
property was bound to "delegateModel.selectedGroup.count".

We will remove the delegate model to manage the selection in the
QAbstractListModel. Since there is no need to expose the selection count
as a bindable property, just update it on drag start.