opengl: use filter input format for "identity"
When an "identity" filter is inserted, we could inject chroma conversion
either in the user filter, or in the "identity" filter.

Since multitexturing is not implemented yet, the "identity" filter is
only implemented for an output formats using only 1 plane (so it can't
output NV12 for example).

Therefore, always inject the chroma conversion in the "identity" filter,
to convert to the input format requested by the OpenGL filter (in
practice, RGBA).