opengl: use common fragment shader for Android
The OpenGL converter for Android used its own minimal fragment shader,
instead of the common one used by all the other converters.

This stemmed from the fact that on each texture update, Android provides
a transform matrix to apply in order to retrieve the correct texture
coordinates.

Adapt the common fragment shader to accept a transform matrix provided
by the converter, and make the Android converter use it.