opengl: use common fragment shader for Android
The OpenGL converter for Android uses its own fragment shader (contrary
to the other converters). This stem from the fact that on each texture
update, Android provides a transform matrix to apply in order to
retrieve the correct texture coordinates.

This is problematic for refactoring the converters to "importers", which
would only initialize what is hardware-specific: they should not deal
with fragment shaders at all.

As a first step, refactor the Android converter to use the "common"
fragment shader generation. That way, the OpenGL vout module will be
responsible for creating fragment shaders (which could then be
refactored/modularized to implement modular filters and renderers).