opengl: explicitly set swizzle_per_tex values
In case the texture format is GL_LUMINANCE, swizzle_per_tex was set to
{NULL, "xa"}. Using NULL instead of an explicit swizzle ("x") was a
small optimization to assign the texel value directly:

    vec4 color0 = texture2D(Texture0, TexCoord0);

instead of using the general code generation:

    colors = texture2D(Texture0, TexCoord0);
    val = colors.x;
    vec4 color0 = vec4(val, val, val, 1);

This was possible because the texture exposes the luminance value in
its three components: (L, L, L, 1).

But this special case will disappear when we will rewrite the chroma
conversion using matrix multiplication directly (instead of computing
its steps separately).

This simplifies the logic of the code, and paves the way to refactor
chroma conversion more easily.