opengl: store fragment shader in sampler
The function opengl_fragment_shader_init() both initializes the
and creates the fragment shader.

To be able to move its initialization outside the renderer, store it in
the sampler instead of returning it. This will also allow to provide the
"extensions" part of the fragment shader as a separate string.