opengl: simplify format initialization
The format provided to vout_display_opengl_New() is passed as a const
pointer to construct the renderer.

The renderer passes it to construct the interop, which copies it and may
change the orientation value in this own copy (only the interop_android
do that). The initial fmt is left unchanged.

The renderer fmt is then initialized from the interop fmt (which is the
same as the initial fmt, expect possibly the orientation).
Then the orientation of the renderer fmt is overriden by the one from
the initial fmt.

As a consequence, the renderer fmt is the same as the initial fmt, and
the whole initialization may be simplified.