opengl: yadif: render first frames correctly
The yadif filter requires 3 frames to operate:
 - next: the frame received during the draw call
 - cur: the frame received during the previous draw call
 - prev: the frame received two draw calls before

This causes a problem for the two first frames.

To mitigate the problem, assign the existing frames to several of the 3
required frames:

 frame | prev  cur  next
   0   | *0*   *0*    0    /* "prev" and "cur" use the "next" frame */
   1   | *0*    0     1    /* "prev" use the "cur" frame" */
   2   |  0     1     2
   3   |  1     2     3
   4   |  2     3     4

That way, for the first frames, the yadif filter will operate as if
successive frames were the same, so only spatial differences will matter
(there will no be temporal differences).